Make update script handle errorsw

+++ _dev/update.sh
@@ -12,21 +12,25 @@ while true ; do
   echo ffmpeg $FFMPEG_LATEST
   echo youtube-dl $YDL_LATEST
 
-  sed -i.bak "s/FFMPEG_VERSION=.*/FFMPEG_VERSION=$FFMPEG_LATEST/g" Dockerfile
-  rm -f Dockerfile.bak
-  if ! git diff --quiet Dockerfile ; then
-    git add Dockerfile
-    git commit -m "Update ffmpeg to $FFMPEG_LATEST"
-  fi
+  if [ "$FFMPEG_LATEST" != "" -a "$YDL_LATEST" != "" ] ; then
 
-  sed -i.bak "s/YDL_VERSION=.*/YDL_VERSION=$YDL_LATEST/g" Dockerfile
-  rm -f Dockerfile.bak
-  if ! git diff --quiet Dockerfile ; then
-    git add Dockerfile
-    git commit -m "Update youtube-dl to $YDL_LATEST"
-  fi
+    sed -i.bak "s/FFMPEG_VERSION=.*/FFMPEG_VERSION=$FFMPEG_LATEST/g" Dockerfile
+    rm -f Dockerfile.bak
+    if ! git diff --quiet Dockerfile ; then
+      git add Dockerfile
+      git commit -m "Update ffmpeg to $FFMPEG_LATEST"
+    fi
 
-  git push origin master
+    sed -i.bak "s/YDL_VERSION=.*/YDL_VERSION=$YDL_LATEST/g" Dockerfile
+    rm -f Dockerfile.bak
+    if ! git diff --quiet Dockerfile ; then
+      git add Dockerfile
+      git commit -m "Update youtube-dl to $YDL_LATEST"
+    fi
+
+    git push origin master
+
+  fi
 
   sleep 86400 # sleep for a day
 done
